application full layout reorganized due to make footer and navbar available on mainLaout and not to use in others layouts
