+++ app/(OtherLayout)/layout.js
@@ -1,0 +1,29 @@
+import { Inter } from "next/font/google";
+import localFont from "next/font/local";
+import AuthProvider from "../AuthProvider";
+import "../globals.css";
+
+const inter = Inter({ subsets: ["latin"] });
+
+const kaleko = localFont({ src: "../../public/fonts/Kaleko105Light.ttf" });
+
+export const metadata = {
+  title: "Avion Realty",
+  description: "Generated by create next app",
+};
+
+export default function SecondaryLayout({ children }) {
+  return (
+    <html lang="en" className="bg-black text-white">
+      <body className={kaleko.className}>
+        <AuthProvider>
+          {" "}
+          <div className="">
+            {/* dynamic content */}
+            <div className="min-h-screen">{children}</div>
+          </div>{" "}
+        </AuthProvider>
+      </body>
+    </html>
+  );
+}
