--- app/login/layout.js
@@ -1,22 +1,0 @@
-import localFont from "next/font/local";
-
-const kaleko = localFont({ src: "../../public/fonts/Kaleko105Light.ttf" });
-
-export const metadata = {
-    title: "Avion Realty",
-    description: "Generated by create next app",
-};
-
-const LoginLayout = ({ children }) => {
-
-    return (
-        <div className="bg-black text-white">
-            <div className={kaleko.className}>
-                {/* dynamic content */}
-                <div className="min-h-screen">{children}</div>
-            </div>
-        </div>
-    );
-};
-
-export default LoginLayout;
